Fix SearchViewControllerTests_testView_EmptyState
